Simplified stdout logger
The TChan-based concurrent log writer appeared unreliable, retaining log
output for long periods, sometimes until the end of the import run. Not
solving the underlying issue here, instead, we get rid of the concurrent
logging business, and instead just guard log writing with a TMVar to
force batches of log lines to be written atomically, and explicitly
hFlush stdout so that log output is produced immediately.