Fixup for #31: don't crash on non-link square brackets
The parser currently only accepts things as shorthand links that
actually look like links, but the previous version would error out when
facing square brackets that are not valid shorthand links. Instead, we
now fail gracefully (`try`), which causes the parser to fall through to
the "plaintext" parser, and that is exactly what we want.