Fix user-finding logic to be more strict
This is important: the /user/?search API returns all users whose email
or username contains the search string, so for example "search=simon"
will produce both simons, as well as any other user with the string
"simon" anywhere in their username or email address. Clearly, this is
not what we want, so the new findUserByEmail function wraps the
search-user API call and applies a post filter to find only those users
whose email address matches exactly. This takes care of previously
encountered inconsistencies, such as finding the wrong users, or failing
due to finding multiple users for a given query.